+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * `TemTemApiRewards.startDate` & `TemTemApiRewards.endDate` are now `DateTime`
 * Removed `TemTemApiPatch.pathInfo`, replaced it with `TemTemTemApiPatch.fixes`, `TemTemApiPatch.improvements` & `TemTemApiPatch.features` properties
 * `TemTemApiPatch.date` is now `DateTime`
+* `TemTemApiSaipark.startDate` & `TemTemApiSaipark.endDate` are now `DateTime`
 
 # 0.1.1+3